remote_cs07/app/requirements.txt
#
awscli==1.16.300
boto3==1.10.36
botocore==1.13.36
boto3-type-annotations==0.3.1
rope==0.14.0
paramiko==2.7.1
#
Flask==1.1.1
Flask-HTTPAuth==3.3.0
Flask-SQLAlchemy==2.4.1
dataset==1.1.2
rope==0.14.0
pycrypto==2.6.1
